Instant Messaging Application
The application takes a username and connects to the server. Once connected, the client starts a receive_message microservice.  This service runs in its own Thread and will constantly receive messages sent from the server and display them within the GUI client application. While this service is running, the chat client allows the user to enter messages within the GUI and sends them to the server. The server then takes the messages and broadcasts them to all users in the chat (connected to server). The server allows private messaging a single user by typing first the @screen_name, a space, and then the private message. Doing this will send the message to only that one user. The user of the chat client has the ability to Connect and Disconnect to/from chat servers and is able to enter the IP address of the server they wish to connect too.