Tic Tac Toe
The form consists of 9 buttons that each player takes turns clicking trying to mark 3 in a row.  Player 1 goes first and will mark a button with an X.  Next, Player 2 takes a turn marking a button with an O.  The winner is the first to get 3 marks in a row, column, or diagonal.  If all 9 buttons are marked and no winner, a "Cats Game" is declared, which is a tie.  After a game (win or tie), the board is reset and starts over with Player 1’s turn.