Student Entry Form
This program is a data entry form. This application allows the user to enter students (two types Student and GradStudent). Radio buttons are used to select the student type while all other information will be entered in Entry widgets. One exception is the student’s major is selected using a Combobox. After all entries are made the user clicks the Save button to add the student object to a list and a Listbox widget for display. The user also has the option to double-click students in the Listbox in order to populate that student’s information into the form for edits.