Calculator
A basic five-function Calculator using the tkinter module that allows addition, subtraction, multiplication, and division of two numbers along with square root of a single number. It allows users to enter numbers ONLY using buttons, only 1 decimal point in a number (two in a number isn’t valid), is able to toggle negative or positive numbers, and is able to recover and not crash when a number is divided by zero.