Employee Management System
This program contains a data list of employees and managers. A manager is an employee as well. However, in addition to managers having all the characteristics of employees, managers also manage employees as subordinates. Therefore, a manager has a list of employees they manage as well as a count of how many subordinates they have. This applications asks the user for details to maintain the list of employees. For each employee, the program asks the user for the name and whether or not the employee is a manager. If the user specifies that the employee is a Manager, the program then asks the user how many subordinates they manage, as well as their names. A manager object is created, and within the manager object, the subordinate employees are created. The program asks the user for the name of the subordinate employees and stores their names in the subordinate employee objects. If the user specifies the employee is not a manager, the program just creates an employee object. All employees (including managers), are added to the data list. Finally, the program prints out a list of all employees. If an employee is a manager, a list of their subordinate employees are also printed.