Grade Calculator for Python Class
This grade calculator calculates your final grade for ASU's CIS 345 course, utilizing the formula specified in the syllabus. The program prompts the user for their total earned points in each weighted grade category (i.e. PE, Assignment, Quiz, etc.).  The data for each category such as earned points, total points that can be earned, weighted percentage of overall grade the category is worth, and the grade or percent you earned out of the weighted percentage for the category is stored and kept track of.